YG-59 🟢feat: free-form 생성
