+++ app/(afterLogin)/createPost/memoForm/page.tsx
@@ -1,27 +1,16 @@
 "use client"
 
 import { useState } from "react"
-import ThumbnailUploader from "../createPost/_components/thumbnailUploader"
-import Overlay from "@/components/commons/overlay"
-import PreviewPostCard from "../createPost/_components/previewPostcard"
-import checkIcon from "@/public/icons/white_check.svg"
-import BackIcon from "@/public/icons/white_arrow-left.svg"
-import FormBtn from "../createPost/_components/form/formBtn"
-import FormInputs from "../createPost/_components/form/formInputs"
-import { QuillEditor } from "../createPost/_components/form/editorQuill"
+import FormBtn from "../_components/form/formBtn"
+import FormInputs from "../_components/form/formInputs"
+import { QuillEditor } from "../_components/form/editorQuill"
 import AddMemoIcon from "@/public/icons/plus-circle.svg"
 import Image from "next/image"
+import UploadOverlay from "../_components/uploadOverlay"
 
 const Page = () => {
     const [isOverlayOpen, setIsOverlayOpen] = useState(false)
-    const [isPreviewVisible, setIsPreviewVisible] = useState(false)
-    const [selectedImage, setSelectedImage] = useState<string | null>(null)
     const [quillEditors, setQuillEditors] = useState<number[]>([])
-
-    const handleComplete = (selectedImage: string | null) => {
-        setSelectedImage(selectedImage)
-        setIsPreviewVisible(true)
-    }
 
     const handleAddMemoClick = () => {
         setQuillEditors([...quillEditors, quillEditors.length])
@@ -35,23 +24,7 @@ const Page = () => {
 
     return (
         <div className="flex flex-col justify-center items-center mb-[205px]">
-            <Overlay
-                isOpen={isOverlayOpen}
-                onClick={() => setIsOverlayOpen(false)}
-                onLeftClick={() => setIsPreviewVisible(false)}
-                leftText={isPreviewVisible ? "다시 선택" : ""}
-                text={isPreviewVisible ? "확인했어요!" : ""}
-                textColor={"text-SYSTEM-white"}
-                imageUrl={isPreviewVisible && checkIcon}
-                leftImageUrl={isPreviewVisible && BackIcon}
-                rounded={"rounded-3xl"}
-            >
-                {!isPreviewVisible ? (
-                    <ThumbnailUploader onComplete={handleComplete} />
-                ) : (
-                    <PreviewPostCard selectedImage={selectedImage} />
-                )}
-            </Overlay>
+            <UploadOverlay isOverlayOpen={isOverlayOpen} setIsOverlayOpen={setIsOverlayOpen} />
             <div className="w-[900px] h-full font-pretendard ">
                 <FormInputs formText={"간단하게 "} />
                 <QuillEditor index={-1} />
