+++ app/globals.css
@@ -24,22 +24,23 @@ body {
 
 /* editor */
 .quill-editor {
-  @apply bg-SYSTEM-white w-[900px] border-0 rounded-xl;
+  @apply bg-SYSTEM-white w-[900px] mt-5 border-0 rounded-xl shadow-lg;
 }
 .quill-editor .ql-snow {
   @apply border-none h-[80px] flex justify-center items-center !important;
 }
 .quill-editor .ql-editor {
-  @apply border-none w-[900px] min-h-[200px];
+  @apply relative border-none w-[900px] min-h-[200px] bg-SYSTEM-white rounded-lg;
 }
 .quill-editor .ql-container  {
-  @apply w-[900px] min-h-[200px] h-max p-2 text-sm text-left bg-SYSTEM-white rounded-xl !important;
+  @apply w-[900px] min-h-[200px] h-max mt-5 p-2 text-sm text-left bg-SYSTEM-white rounded-xl !important;
 }
 
-/* memo-form editor style */
-.memo-editor .ql-container {
-  @apply mt-5 !important;
+/* free-form editor style */
+.free-editor .ql-container {
+  @apply h-[1060px] !important;
 }
+
 
 @layer utilities {
   .text-balance {
